parse insert with functions
Allows functions with parenthesis to appear in an INSERT VALUES list.
Also added tests for that case. This patch adds a dependency on
Text::Balanced, which is pure Perl and works with older Perl versions. I
considered Regexp::Common, but figuring out a regexp that worked with
old & new perl's both was more work than I cared for.

ps. The original code allowed multiple value lists, eg "insert into foo
values ('a',1),('b',2),('c',3)" so I kept that functional and also added
a test for that.