SQL::Statement::Operation - fix RT#99349 (bind variables are not bound to placeholders in the correct order)
Bind variables are not bound in the correct order to the placeholders in a
'where'-clause of a SQL statement, if some parts of the clause need not to
be executed (e.g. OR after a true condition).

Add appropriate test for issue in t/02execute.t

Found and fix provided by vulpeculus@gmx.de

Signed-off-by: Jens Rehsack <sno@netbsd.org>